+++ src/components/ProfileModal/components/pages/index.ts
@@ -1,4 +1,5 @@
 export { default as SettingsPage } from './SettingsPage';
+export { default as AchievementsPage } from './AchievementsPage';
 export { default as ProfilePage } from './ProfilePage';
 export { default as ThemePage } from './ThemePage/ThemePage';
 export type { ThemePageRef } from './ThemePage';